+++ libs/api/TcpServer.cpp
@@ -10,7 +10,7 @@ static void echo_alloc(uv_handle_t* handle,
 	size_t suggested_size,
 	uv_buf_t* buf) {
 	buf->base = (char*)malloc(suggested_size);
-	buf->len = suggested_size;
+	buf->len = (ULONG)suggested_size;
 }
 
 static void close_cb(uv_handle_t* handle) {
@@ -77,7 +77,7 @@ void TcpServer::_ConnectionCB(uv_stream_t* remote, int status)
 void TcpServer::ReadCB(uv_stream_t * stream, ssize_t nread, const uv_buf_t * buf)
 {
 	if (nread < 0) {
-		RLOG("read_cb error: %s\n", uv_err_name(nread));
+		RLOG("read_cb error: %s\n", uv_err_name((int)nread));
 		assert(nread == UV_ECONNRESET || nread == UV_EOF);
 		uv_close((uv_handle_t*)stream, close_cb);
 	}
@@ -155,12 +155,12 @@ int TcpServer::Write(void * buf, size_t len)
 	}
 
 	uv_write_t* write_req = new uv_write_t();
-	uv_buf_t uvbuf = uv_buf_init((char*)buf, len);
+	uv_buf_t uvbuf = uv_buf_init((char*)buf, (unsigned int)len);
 	int rc = uv_write(write_req,
 		(uv_stream_t*)m_tcp_client,
 		&uvbuf,
 		1,
 		WriteCB);
-	RLOG("%s Send len %ld returns %d\n", __FUNCTION__, len, rc);
+	RLOG("%s Send len %llu returns %d\n", __FUNCTION__, len, rc);
 	return rc;
 }
